Conflicts:
	src/main/java/com/canfer/app/model/ComprobanteFiscal.java
	src/main/java/com/canfer/app/rest/FNCRestController.java

+++ src/main/java/com/canfer/app/model/DocumentosNacionalesActions.java
@@ -288,9 +288,17 @@ public class DocumentosNacionalesActions extends ModuleActions {
 		if (value.isPresent()) {
 			
 			comprobanteFiscal = value.get();
-			
+
 			if(comprobanteFiscal instanceof ComplementoPago) {
 				clearComplemento((ComplementoPago) comprobanteFiscal);
+			}
+			
+			if (comprobanteFiscal instanceof Factura
+					&& ((Factura) comprobanteFiscal).getComplemento() != null) {
+				
+					return false;
+					
+				
 			}
 			
 			comprobanteFiscal.delete();
@@ -322,13 +330,21 @@ public class DocumentosNacionalesActions extends ModuleActions {
 					clearComplemento((ComplementoPago) comprobante);
 				}
 				
-				comprobante.delete();
+				if (comprobante instanceof Factura
+						&& ((Factura) comprobante).getComplemento() != null) {
+						
+					
+				} else {
 				
-				Log.activity("Se ha eliminado el documento fiscal No. " + comprobante.getIdNumSap() + "." , comprobante.getEmpresaNombre(), "DELETE");
+					comprobante.delete();
+					superRepo.delete(comprobante);
+					
+					Log.activity("Se ha eliminado el documento fiscal No. " + comprobante.getIdNumSap() + "." , comprobante.getEmpresaNombre(), "DELETE");
+				
+				}
 				
 				});
 			
-			superRepo.deleteAllComprobante(comprobantes);
 			
 			return true;
 			
@@ -361,6 +377,7 @@ public class DocumentosNacionalesActions extends ModuleActions {
 	public boolean updateCfdInformation(ComprobanteFiscalDTO documento) {
 		
 		Optional<Proveedor> proveedor = null;
+		ComprobanteFiscal comprobanteUpdate;
 		Optional<ComprobanteFiscal> comprobante = superRepo.findComprobanteById(documento.getIdComprobanteFiscal());
 		
 		if (comprobante.isPresent()) {
@@ -373,17 +390,26 @@ public class DocumentosNacionalesActions extends ModuleActions {
 			
 			try {
 				
-				if(comprobante.get().actualizar(documento, proveedor)) {
+				int choice = comprobante.get().actualizar(documento, proveedor); 
+				
+				if(choice == 1) {
 					
-					emailSender.sendEmailUpdateDoc(comprobante.get());
-					superRepo.save(comprobante.get());
-					
+					comprobanteUpdate = superRepo.save(comprobante.get());
+					emailSender.sendEmailUpdateDoc(comprobanteUpdate);
 					return true;
 					
-				} else {
+				} else if (choice == 2 || choice == 3) {
 					
+					comprobanteUpdate = superRepo.save(comprobante.get());
+					emailSender.sendEmailNewDoc(comprobanteUpdate);
+					return true;
+					
+				}else {
+				
+					superRepo.save(comprobante.get());
 					return false;
 				}
+				
 				
 			} catch (Exception e) {
